+++ packages/core/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.9.0
+
+#### 🎉 Release
+
+Added a new type of output to concurrently render progress bars in a handful of different styles.
+
+#### 🚀 New
+
+- Added `ProgressOutput` class.
+- Added `Output#onStart`, `onComplete`, `onFirst`, `onLast` lifecycle methods for sub-classes.
+- Added `Reporter#createProgressOutput` to instantiate a concurrent progress bar output.
+
 # 1.8.0 - 2019-02-06
 
 #### 🚀 New